ARTEMIS-4372 Implement Pico-cli and script auto-complete ARTEMIS-4375 Implement artemis shell using JLine3 integrated with auto-completion from picocli
This commit involves two JIRAs. One is adding PicoCLI and the next is Using JLine3 and implement a shell.
I have tried to keep these commits separate but these changes became interdependent hence the two JIRAs are squashed in this commit.

+++ artemis-cli/src/main/java/org/apache/activemq/artemis/cli/commands/messages/ConsumerThread.java
@@ -33,7 +33,7 @@ import java.util.concurrent.CountDownLatch;
 
 public class ConsumerThread extends Thread {
 
-   int messageCount = 1000;
+   long messageCount = 1000;
    int receiveTimeOut = 3000;
    Destination destination;
    Session session;
@@ -259,7 +259,7 @@ public class ConsumerThread extends Thread {
       return this;
    }
 
-   public ConsumerThread setMessageCount(int messageCount) {
+   public ConsumerThread setMessageCount(long messageCount) {
       this.messageCount = messageCount;
       return this;
    }
@@ -278,7 +278,7 @@ public class ConsumerThread extends Thread {
       return this;
    }
 
-   public int getMessageCount() {
+   public long getMessageCount() {
       return messageCount;
    }
